+++ web/index.html
@@ -2,7 +2,6 @@
 <html>
   <head>
     <link rel="stylesheet" href="style/style.css">
-    <link rel="stylesheet" href="style/createProject.css">
     <link rel="stylesheet" href="https://cdn.rawgit.com/twbs/bootstrap/v4-dev/dist/css/bootstrap.css">
     <link rel="stylesheet" href="style/fonts/font-awesome/css/font-awesome.min.css">
   </head>
