+++ web/style/style.css
@@ -92,3 +92,149 @@ color: white;
 .inner-content {
   position: relative;
 }
+
+/* createProjectByPattern.js and createProjectView.js */
+
+body{
+  width: 1440px;
+  height: 1024px;
+  background-color: #ffffff;
+}
+
+.content{
+
+
+}
+
+.header{
+    width: 388px;
+    height: 41px;
+    font-family: AvenirNext;
+    font-size: 30px;
+    color: #000000;
+    margin-left: 657px;
+    margin-top:262px;
+}
+
+.text{
+  width: 30px;
+  height: 41px;
+  font-family: AvenirNext;
+  font-size: 30px;
+  color: #000000;
+  margin-left: 836px;
+  margin-top:46px;
+}
+
+.enterUrl{
+  width: 548px;
+  height: 25px;
+  font-family: AvenirNext;
+  font-size: 18px;
+  font-weight: 200;
+  color: #676767;
+  margin-left: 550px;
+  margin-top:46px;
+
+
+}
+
+.submit{
+  width: 250px;
+  height: 43px;
+  border-radius: 5px;
+  background-color: #ff5500;
+  border: solid 1px #ff5500;
+  font-family: AvenirNext;
+  font-size: 18px;
+  margin-left: 725px;
+  margin-top:46px;
+
+}
+
+.headerCreation{
+  width: 316px;
+  height: 41px;
+  font-family: AvenirNext;
+  font-size: 30px;
+  color: #000000;
+  margin-top: 79px;
+  margin-left: 293px;
+}
+
+.caption{
+  width: 108px;
+  height: 19px;
+  font-family: AvenirNext;
+  font-size: 14px;
+  color: #999999;
+  margin-left: 293px;
+
+}
+
+.enterValue{
+  width: 841px;
+  height: 32px;
+  background-color: #ffffff;
+  box-shadow: inset 0 1px 0 0 #ececec;
+  border: solid 1px #cccccc;
+  margin-left: 290px;
+
+}
+
+.dropdown{
+  width: 400px;
+  height: 32px;
+  background-color: #ffffff;
+  box-shadow: inset 0 1px 0 0 #ececec;
+  border: solid 1px #cccccc;
+  margin-left: 290px;
+}
+
+.enterDescription{
+  width: 841px;
+  height: 105px;
+  background-color: #ffffff;
+  box-shadow: inset 0 1px 0 0 #ececec;
+  border: solid 1px #cccccc;
+  margin-left: 290px;
+}
+
+.tags{
+  width: 841px;
+  height: 32px;
+  background-color: #ffffff;
+  box-shadow: inset 0 1px 0 0 #ececec;
+  border: solid 1px #cccccc;
+  margin-left: 290px;
+}
+
+.buttons{
+  margin-top: 192px;
+
+}
+
+.saveButton{
+  width: 119px;
+  height: 43px;
+  border-radius: 5px;
+  background-color: #333333;
+  border: solid 1px #333333;
+  margin-left: 296px;
+  font-family: HelveticaNeue;
+  font-size: 18px;
+  text-align: center;
+  color: #ffffff;
+}
+
+.cancelButton{
+  width: 99px;
+  height: 43px;
+  border-radius: 5px;
+  border: solid 1px #cccccc;
+  margin-left: 37px;
+  font-family: HelveticaNeue;
+  font-size: 18px;
+  text-align: center;
+  color: #000000;
+}
